+++ RuntimeLib/World.c
@@ -261,7 +261,7 @@ Void RTWorldSpawnMob(
         Mob->ID,
         X,
         Y,
-        Mob->Attributes.Values[RUNTIME_ATTRIBUTE_MOVEMENT_SPEED],
+        (Int32)Mob->Attributes.Values[RUNTIME_ATTRIBUTE_MOVEMENT_SPEED],
         RUNTIME_WORLD_TILE_WALL | RUNTIME_WORLD_TILE_TOWN
     );
     RTWorldTileIncreaseMobCount(Runtime, WorldContext, Mob->Movement.PositionTile.X, Mob->Movement.PositionTile.Y);
@@ -722,7 +722,7 @@ Void RTWorldRespawnMob(
         Mob->ID,
         X,
         Y,
-        Mob->Attributes.Values[RUNTIME_ATTRIBUTE_MOVEMENT_SPEED],
+        (Int32)Mob->Attributes.Values[RUNTIME_ATTRIBUTE_MOVEMENT_SPEED],
         CollisionMask
     );
     RTWorldTileIncreaseMobCount(Runtime, WorldContext, Mob->Movement.PositionTile.X, Mob->Movement.PositionTile.Y);
