+++ WorldSvr/OnChat.c
@@ -13,7 +13,7 @@ CLIENT_PROCEDURE_BINDING(MESSAGE_NEARBY) {
 
 	PacketBufferRef PacketBuffer = SocketGetNextPacketBuffer(Socket);
 	S2C_DATA_MESSAGE_NEARBY* Notification = PacketBufferInit(PacketBuffer, S2C, MESSAGE_NEARBY);
-	Notification->CharacterIndex = Character->CharacterIndex;
+	Notification->CharacterIndex = (UInt32)Character->CharacterIndex;
 	Notification->Nation = Character->Data.StyleInfo.Nation;
 	Notification->PayloadLength = Packet->PayloadLength;
 	PacketBufferAppendCopy(PacketBuffer, &Packet->Payload[0], Packet->PayloadLength);
